requirements.txt
fastapi
uvicorn
python-dotenv
langchain>=0.1.0
langchain-huggingface>=0.0.6
langchain-community>=0.0.10
transformers>=4.36.0
torch>=2.1.0
pydantic>=2.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
accelerate>=0.25.0
numpy>=1.24.0
